+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-3](https://github.com/OS2web/os2web_key/pull/3)
+  Added Azure Key Vault certificate provider
 * [PR-2](https://github.com/OS2web/os2web_key/pull/2)
   Updated documentation
 * [PR-1](https://github.com/OS2web/os2web_key/pull/1)